fix(cluster): do not use channel created for a different address
When only using IP to find a channel pool, we observed that a channel pool
created for a node is re-used by another node which got its old ip. This
could lead to unexpected situations, where messages are sent to wrong node.

To fix this, we now use both address and ip to find a channel pool. In a
setup where restarts and ip re-assignment is common, it is safer this way.

(cherry picked from commit 943f01473bf47869dab3b6d20769777d7a58fbb8)