feat(engine): pass job activation behavior to processors
Pass the BpmnJobActivationBehavior instance to all the Processor classes that change the
job state. Any Processor class that mutates the job state to ACTIVATABLE needs to have access
to the BpmnJobActivationBehavior class so that the job can be immediately pushed to the job
worker. With later commits, we plan to refactor the long-polling notification, so that it is
done as a side effect through the BpmnJobActivationBehavior.