fix(engine): use static error code without evaluating expression
Fix for https://github.com/camunda/zeebe/issues/12326

When we transform the error and it's a static value, we would still evaluate the expression. This could result in
either a STRING or a NUMBER. As NUMBER wasn't a case that was implemented it would cause processes to be unable to
 find catch events.

If it's a static value we can just take the error code from the error event, like we did before we support FEEL
expressions.