fix(engine): use static escalation code without evaluating expression
Fix for https://github.com/camunda/zeebe/issues/12326

When we transform the escalation and it's a static value, we would still evaluate the expression. This could result in
either a STRING or a NUMBER. As NUMBER wasn't a case that was implemented it would cause processes to be unable to find
 catch events.

If it's a static value we can just take the escalation code from the escalation event.

(cherry picked from commit efc4a56c111e0a5f1560c8c771b1d9018c654444)