fix(broker): allow disabling flush alone
Allow users to configure disabling flush by itself without having to
specify a delay. Previously this was not possible because the config is
a record, and we enforced that the delay must not be null.

The fix is to provide a default value if the delay time is null, if only
to continue allowing the usage of the config without worrying about null
values.

Unfortunately it's not possible to use a second constructor to provide a
null value, primarily because it would require us annotating it with
`@ConstructorBinding`, which is part of the `spring-boot` library, and
so we'd have to add a whole dependency just for that. A simple null
check is good enough in this case.