feat(transport): aggregate logically equivalent client stream into one server stream
Client streams with same streamType and metadata are mapped to a single stream that is registered
with the servers. Payloads pushed to this server stream will be distributed to one of the registered
client stream. Currently, the client stream is picked randomly. Later, we can employ a better strategy
to chose the client stream to push the payload.