fix(transport): do not sent add stream request unnecessarily
If two client streams were added around same time, it is possible that both will
trigger sending add request to the servers. This is ok as the request is idempotent,
but is unnecessay overhead. This commit fixes it by having an explicit INITIAL and OPEN state
in the aggregated client. The request is sent only once when it is the initial state.