refactor(gateway): allow waiting for brokerclient startup
Previously, brokerClient does not wait until all the subcomponents are started. This caused a flaky
test because the topology set by the BrokerClientTest is overwritten by BrokerTopologyManagerImpl.
To fix it brokerClient returns a set of future for all the services it starts. The caller can wait
on this future.

Ideally, it should return one future. But BrokerClient::start is called from actors and non-actors.
So there is no common way to aggregate multiple actorfutures into one. If the caller is actor, we can
use ActorFutureCollector to get a composite future. If the caller is non-actor this is not possible,
instead we can simply call join on each of the future, which is not allowed if the caller is the actor.
So the current solution is just an ugly workaround to return a collection of future. The caller can decide
how to wait for the startup completions.