feat: introduce cache for process version
In order to avoid unnecessary RocksDB access,
we introduce an cache for the process version.

The version is likely to not change too often,
which will cause it to move to a low layer in
RocksDB, which might make access slow in the hot
path.

(cherry picked from commit 743bec2ca5c74372c3d0d359106d06396c41b008)