fix: use larger fixed length prefix extractor
The prefix extractor is used during iterator, if more bytes can be
considered during seek and iteration it makes the internal search more
performant.

If we iterate we normally do this via a prefix, mostly a long (e.g.
scopeKey) which needs to be considered. The column family itself is
already a long, which means if we use another long in the fixed
extractor we can considere the addition key.