fix: handle failure to resolve address more gracefully
When a cluster member address can't be resolved, for example while the
node is offline for a rolling update or because the dns resolver is not
available, we used to throw `IllegalStateException`. The gateway's
`GrpcErrorMapper` treated those as internal, unexpected errors and
responded as such. We also logged this as errors.

By treating dns resolution failure as `ConnectException`, the gateway
handles this more gracefully and responds to the client with an
availability error so that clients can eventually retry the request.

Retryable requests made via `BrokerClient` now also retry on DNS
resolution errors instead of failing immediately.