feat(transport): add client stream metrics
Adds new metrics for client streams:

- the count of known servers
- the count of aggregated streams
- the count of clients per aggregated streams
- the count of successful pushes
- the count of failed pushes

The metrics are added as an interface, and are also tested. The
implementations themselves can then decide how to best represent them
(e.g. the count of clients per streams can be an histogram).

I opted to add more metrics for now, and we can always drop them while
we do E2E tests if we don't think they will be useful long term.