refactor: change all time durations to interval variable
In order to unify range vectors and time duration we introduced a new
variable. This allows that all panels use the same time duration, and
are comparable. The interval variable defaults to $__rate_interval,
which is provided by Grafana. It estimates based on the prometheus
scrape interval a good sensible value for the time range, which should
be at least contain four data point to make good results for the rate.