feat(engine): verify we can write a command with a specific length
As part of the TerminateProcessInstanceBatchProcessor we want to check if are able to write the termination of the
next child in the batch. This commit exposes a method on the CommandWriter which allows us the check this.

(cherry picked from commit c7c4f2780c631e0eafb77523d21b0421e352e24c)