feat(engine): apply a visitor for each child of an element instance
Introduces the forEachChild method to the ElementInstanceState. This method is used to apply a given
function for each of the children of the parent element instance.
The method takes a startAtKey to indicate which child the iteration should start at.
This will be used for batch termination to prevent sending duplicate terminate commands.

(cherry picked from commit 0e4cb4a5593c3beb055b506f79b587a13fd66cb7)