feat(engine): implement TerminateProcessInstanceBatchProcessor
Adds the implementation to the processor. The processor will iterate over the children of the parent.

It will check if it is able to write this command and a follow-up batch command.
If possible it will write the TERMINATE command for the child.
It returns true to indicate we should continue with the next child.

If it is unable to write both commands, we only write the follow-up batch command.
We set the index of this command to the element instance key of this child.
This is to indicate that we should start at this child in the next batch.
It returns false to indicate we should stop here, and not proceed with the next child.

(cherry picked from commit 3954ef3e2071d4fc4945271644a3ec81f11455e2)