refactor(transport): add ClientStreamService iface
Adds a new `ClientStreamService` interface as the main entry point to
set up the remote stream client side. Current implementation is the
`ClientStreamServiceImpl`, which is also the implementation for the
`ClientStreamer`. This may change in the future to separate
responsibilities.

This is necessary to avoid exposing the complete actor to the outside
world, and allow providing an adapter to add/remove servers.