docs: document flaky test handling
Add a guide on what to do when a flaky test is encountered. The guide is included in the ci.md
section of our developer docs. This section is also linked inside our CONTRIBUTING.md guide to
ensure that any contributors are aware of what to do in such an occurrence.