feat: introduce new interval variable
New interval variable allows to be used for range vectors in order to change rate function dynamically.

Defaults to rate_interval, which is provided by grafana (but needs to be configured).
It estimates a good time duration based on the prometheus scrape
interval, where at least four data points are included for the range
vectors.

Since it might be necessary to configure it in grafana (it defaults to
30s?) it makes sense to provide some other possible values, which can be
used by customers or us in SaaS where this configuration might be
different or is not applicable.