test(engine): fix terminate tests
Fixes existing tests to adhere to batch termination

BoundaryEventTest:
This test asserts over an exact sequence of record. The terminate batch command needs to be inserted in this sequence.

ProcessInstanceCommandRejectionTest:
Because of the batch command no seconds terminate command is written anymore.
We need another setup to write 2 terminate commands.