fix: reset blacklist metric
On restart/recovery we need to restore the blacklist metric
in order to report it always. Otherwise we will have gaps in the reports
and it is not reliable, due to restarts and reseting of counters.

Change counter to gauge, such it can be reset/set to a specific value
based on the state.

(cherry picked from commit ffb030f3428eb21d8df61cd5edf10fde77550cac)