fix: support listing up to 65k backups instead of 255
Switches to a new encoding that uses 16 bits instead of 8 to count
entries.
Rather than changing the generic `groupSizeEncoding` which is used in
other places too, this introduces a second `largeGroupSizeEncoding`.
This is mostly to ensure backwards compatability and to save some bits
where they are not needed.

(cherry picked from commit 7c4254b9063436ad318be1186d43cf394c004458)