fix(backup): do not take backup when one already exists
Previously, we were only checking for backups taken by this node.
However, there could be backups taken by another node. For example, after
restore the checkpoint record is processed again, but it must not re-take
the backup. To prevent that, we should check for all backups with the given
backup id.