fix: use __rate_interval everywhere
Replace the interval variable with the rate_interval, we expect it should work most of the times without issues. If not the user needs to configure grafana correctly, via setting the prometheus scrape interval in the config.

Remove the interval variable from the panel title, in order to avoid confusion.