wip: first approach to introduce caches in CF
Not that easy since we have several different key types, and in the ZeebeDB mostly they are generic, so we can align on a common sense which is the byte array or introduce some method to the DbKey interface, to identify the key like hashcode etc.

It would be easier if we refactor the API to allow direct types/objects see https://github.com/camunda/zeebe/issues/12595 

In POC i tried it with using long hashcode

LRU Cache from agrona not that useful, used later the Cache impl (which is a map with capacity etc.)

Run again into the issue of reusing objects, especially values. The cache was corrupted by this, when storing just the values at somepoint all values have changed to the same value with different keys….

Caching is not that straightforward and lot of traps to fall into

Two layer caches, one for the current transaction and one for committed data?

Dont forget to delete data from cache…