feat(engine): create ActivateProcessInstanceBatchProcessor
Adds the ActivateProcessInstanceBatchProcessor and registers it with the Engine.

This processor is responsible to handle the ProcessInstanceBatch.ACTIVATE commands.
It is used for activating child element instance of a multi instance element in a batch-like manner.
The index in the command is used to indicate how much children need to be activated.
If the amount of children will exceed the max message size a followup batch command is written.
The index of the followup command is decremented for each of the activate commands that have been send.

(cherry picked from commit ddd34ad7f631ac7f7a245bd8c006a75d3a11d2ea)