feat(engine): activate multi instance children in batches
Instead of writing all the ACTIVATE commands for all children at once, we now use the
ProcessInstanceBatch.ACTIVATE command instead.

This command takes an index, which in this case is the amount of children that an ACTIVATE command will be written for.

(cherry picked from commit 65988a2fd5543b9890906985e68ba1b171304cec)