refactor(broker): add stateless job stream error handler
Adds a stateless job stream specific implementation of the
`RemoteStreamErrorHandler`. This error handler delegates to a ZPA
implemented `JobStreamErrorHandler` which will build a task result,
allowing them to append follow up command using the tools they're
familiar with.

In order to write per partition, the implementation is a partition
listener as well, and an actor, as it may be called from different
threads.