ci: don't compile and test the java client with JDK 8
This was there to ensure that we don't accidentally break compatibility
with Java 8.
As it turns out, we already make use of the `--release` feature in javac
which ensures that we don't accidentally start using newer APIs.
For example, adding a line such as `Set.of(1, 2, 3).stream().toList();`
to the java client results in a failure to compile.

See also https://openjdk.org/jeps/247

(cherry picked from commit d1d4128d2ea099fcd2b991ba3c333fbac18b9874)