feat: add record version to record metadata
The record version will allow us to differentiate between between
older and newer records.

When an event is applied, then it should always be applied in the same
way. Even when that application was faulty and led to a bug. We can only
resolve the bug for future events (not yet written events). In order to
differentiate between the application of an old event and a new event,
we need some way to indicate their difference on the log.

Initially, we thought to use the brokerVersion for this, but it is hard
to write code referring to major.minor.patch versions that have not yet
been released. It is much simpler to refer to v1 or v2 of a record.

This is why a new record field is necessary, to record the record's
version on the log.

Although initially only used for events, it could become useful for
commands or command rejections as well. This is why it is called
recordVersion instead of eventVersion.