refactor(stream-platform): extract metadata parameter
We need to add a new paramater to this method (recordVersion), further
increasing the already long list of parameters. This recordVersion will
exist on the RecordMetadata. It makes sense to just extract the creation
of the RecordMetadata to a parameter so the recordVersion can be set at
the callsite.

This is only the first of several such refactorings in attempt to first
make the hard change easy, and then make the easy change.