refactor(stream-platform): extract metadata parameter
This is another refactoring pulling the RecordMetadata out of methods
that get provided all the relevant parts that make up the metadata as
parameters. This will make a future change to add recordVersion to the
metadata easier. Instead of further expanding the already long list of
parameters, we'll just set the value on the metadata that is passed in.

An additional change seemed necessary here for the RecordBatchTest. It
did many calls the changed method and would need to rebuild the metadata
in each call. However, for most of these tests the actual data in the
metadata doesn't matter. Therefore I've extracted it into a constant
that is build up out of other constants that can be used in the one test
case that cares about these values.