refactor: extract metadata parameter for appendRecord
We want to introduce a new recordVersion property to records. We'd need
to expand the already long list of parameters of this method with
another entry. To avoid this we can extract the metadata to a parameter.
In addition to making it easier to add the recordVersion property (it
will be part of the metadata), it also cleans up the callsites by being
expressive about the data provided to the method call.

An alternative for this change was to use overloading. This could help
to avoid making changes in the backup, engine, and stream-platform
modules simultaneously. However, I believe this change is benefitial to
all as it helps with the code readability.