test(engine): verify event appliers registrations
Event appliers have always been an implementation detail, and still are.
But now that we want to have versioned event appliers, it makes sense to
add some tests that show that we can apply state using the registered
applier that belongs to that event. That includes testing that the
recordVersion matters in the application of events and the registration
of appliers.

Typically, I'm not a fan of using Mockito.verify but it seems like the
most reasonable way to test that the mapping can be adjusted by
registering appliers. It can be considered that this tests the behavior
of this class, i.e. it's purpose is to provide a way to use the correct
event applier for an event.