feat(journal): update lastIndex and position before creating next segment
lastIndex and lastPosition is updated only when moving to next segment. This
is to keep it simple, if we keep updating it for every append it is difficult
to keep them in sync in disk because the blocks might be flushed in different order.
It is ok to not have it set for the current segment. After restart, the writer has
to scan only one segment. Besides it is better to scan and built the index of the
last segment as the consumers of journal is most likely to seek to this segment.