fix(journal): avoid gaps in logs due to reset
Avoid gaps in the log when resetting the log by deleting the segments in
reverse order. This may cause data loss, but that's acceptable in the
context of a reset operation. Deleting segments in reverse order will
ensure we have no gaps between the remaining snapshot and segments
should the operation fail midway.

Note that this does not make persisting a new snapshot/reseting the log
atomic unfortunately.

Unfortunately tests were also not ideal, and require heavy use of mocks.
It's also not possible to write integration test without major
refactoring. As this is a quick fix, the hope is that the proper
solution to make the Raft install operation atomic will redeem this.