fix: don't attach event subprocess to multi-instance body
The event sub process should only be attached to the inner instance of
the multi-instance marked element. Not to the body.

The event sub process should be triggered for a specific instance of a
multi-instance marked element only. It only exists in the flow scope of
the inner instance, not in the flow scope of the multi-instance body.

There was a bug https://github.com/camunda/zeebe/issues/11578 with this
as the root problem. It surfaced because the event sub process's start
event had a correlation key expression that referred to the input
element as variable. That variable had not yet been created when we
subscribed the multi-instance body to the event sub process. We should
not have attempted to subscribe the multi-instance body to the event sub
process. Instead, each inner instance of the multi-instance should be
subscribed to its own instance of the event sub process.

See the previous commit for a test case with that scenario.