test: trigger event subprocess in multi-instance
An event sub process existing inside a multi-instance sub process should
exist for each inner instance of the multi-instance. Each event sub
process should be able to be triggered individually.

There was a bug https://github.com/camunda/zeebe/issues/11578, where the
event sub process's message start event could not create a message
subscription because the correlation key expression referred to the
input element. The input element exists on each inner instance, but the
engine attempted to subscribe to the event sub process before activating
any of the inner instances. In turn, this also means that the event sub
processes couldn't be triggered.

Instead of testing that we can subscribe to each event sub process, it
makes more sense to also test that we can trigger each individually.
This tests the reason why we want to be able to subscribe in the first
place.