fix(protocol-impl): set a default for recordVersion 0
Although, we carefully considered the SBE nullValue of the uint16 type
of the recordVersion in b826f7f, it seems that a bug slipped through the
cracks.

I first discovered this bug, because the UpdateTests (NoSnapshotTest,
etc) were failing on my branch korthout-12764-version-event-appliers. I
did not understand how my changes on that branch could be breaking
backward compatibility.

Then, I bisected the commits and discovered that 9c1a7d2a introduced a
bug. That commit is where we switch event applying over to versioned
event appliers. Since all records are assumed to be on the default
version (1), it was expected that all events could be applied
automatically by version 1 of each event applier. However, all records
written in 8.2.0 were considered to have recordVersion `0` instead.

Our tests in JsonSerializableToJsonTest did not catch this, because it
writes the empty record using a new engine that already knows the new
protocol. And our tests in UpdateTests did not catch this before,
because the new recordVersion was not used for any real processing yet.

On further analysis, it seems that we mis-interpreted the implementation
of RecordMetadataDecoder.recordVersion(). We expected it to return 65535
(the nullValue of recordVersion) for records where this field is
missing. However, for whatever reason, in that test the method returns 0
instead.

Why this is, is unclear to me. It reads to me as if the wrap passes the
version in the header along (which I expect to be 2), but somehow it is
3 or higher.

In any case, we'll have to make sure to set a default recordVersion in
case the decoder returns 0. Whether or not we understand why, it cannot
hurt to do so, and more importantly, it resolves the problem.