fix(stream-platform): verify isCommandDistributed only checks commands
A TypedRecord could be an event, command or command rejection. The name
of this method suggests that it checks if a command is distributed. It
doesn't relate to events or command rejections. We should make sure we
only check commands in this method. If it's not a command we can
consider the record to not be distributed.

Signed-off-by: Remco Westerhoud <remco@westerhoud.nl>