refactor: rename persisted command distribution
The CommandValueAndValueTypeWrapper was explained to be used to store
the value and the value type together. However, the real reason it
exists is to allow storing the details of a CommandDistributionRecord.

We cannot store a CommandDistributionRecord in a ColumnFamily directly,
because CommandDistributionRecord does not implement DbValue.

In a future step, I'll need to persist more details about the command
distribution in the state along side the value and the value type. So,
it only makes sense to rename it now to PersistedCommandDistribution.

I've removed the JavaDoc, as I don't think this class needs further
explanation with the new name.