fix(protocol): support json serialization of unknown intent
Apparantly, we were unable to serialize the UNKNOWN intent, which is
used when we don't know the ValueType, or don't recognize the intent's
value, or when we used VAL_NUL.

I discovered this, because JsonSerializableToJsonTest failed to
serialize the empty CommandDistributionRecord when Intent was added.

The error highlighted the problem:

java.lang.RuntimeException: com.fasterxml.jackson.databind.exc.InvalidDefinitionException:
 No serializer found for class io.camunda.zeebe.protocol.record.intent.Intent$1 and no
 properties discovered to create BeanSerializer (to avoid exception, disable
 SerializationFeature.FAIL_ON_EMPTY_BEANS) (through reference chain:
 io.camunda.zeebe.protocol.impl.record.value.distribution.CommandDistributionRecord["intent"])

The anonymous class Intent$1 cannot be serialized. Typically this
happens when the class doesn't have getters. So I added getIntent() to
it, but then the result was an object with intent as field:
```
{"intent":{"intent":255},"partitionId":1,"valueType":"NULL_VAL","commandValue":null}
```

I ended up replacing the UNKNOWN intent with a static enum implementing
Intent in the same way as before. (static is left out because it's an
inner enum, and these are always static).