feat: write intent in command distribution records
Now, we need to make sure that we actually set the intent with a value
when we write the STARTED event, so it can be stored in the state in the
event applier.

We don't need to write the intent for the DISTRIBUTING event. We have no
access to it anymore in the ACKNOWLEDGE command, so it also is not
available in the ACKNOWLEDGED and FINISHED events. So it would simply be
distracting in the DISTRIBUTING event, as it does not help track what
happens on the log.