refactor: add intent to persisted command distribution
We want to store the intent in the state, so we can read it later in the
Command Redistributor.

The wrap method is only called from the DbDistributionState when a
command distribution is added to the state, which is only done as part
of the STARTED event applier. The STARTED event has the `intent` filled
in already.