feat(engine): allow accessing all flow elements of an ExecutableProcess
When trying to find a straight-through processing loop it is not
necessary to check all elements. We only have to verify the elements
that are straight-through processing. Adding this method allows us to
access all elements of the process, which we can use to later filter for
 specific element types.

(cherry picked from commit 8a7990e12fcc8e9ade507056bef9870b826d0111)