feat(engine): detect loops containing subprocesses
This adds a few more straight through processing element types. This is
used to detect loops which contain a subprocess in which there are also
only straight through tasks.

It also slightly alters the way the algorithm runs. When we encounter a
subprocess we can't just check the outgoing sequence flows. Instead we
must check the path inside of the subprocess.

This also means that when we reached the end of the subprocess we must
continue back at the flowscope.

(cherry picked from commit 3f495f4c7f23d42f9e8576fbc6b13e06bae9033c)