feat(engine): deal with stackoverflows
As we use recursion there is a chance of running into a
StackOverflowError. The chances of this are very very slim. I tried to
cause it myself but even with a chain of 2000 tasks I wasn't able to.
It's highly unlikely any user will do this, but better be safe than
sorry.

I have tested this manually by throwing a StackOverflowError. I couldn't
 write a performant test for this unfortunately.

(cherry picked from commit 6020676ee43f2996164cb582f242fdf1ee09d559)