test(broker): add true/false distinction between tests
The environment and the config file must have a different value. The
test is supposed to verify that we the value of the config file is
overwritten by the environment. If they have the same value there is no
way to know if it came from the config file, or from the environment.