test(engine): verify call activity calling itself is rejected
Call Activities calling itself is one of the scenarios we've seen
causing a straight-through loop. This is only a problem if the call
activity calls it's own process without any wait state in between. E.g.
start event > undefined task > call activity > end event. If the call
activity here calls itself, we're in a straight-through loop.

If there is a wait state in between, e.g.:
start event > user task > call activity > event. We do not detect a
straight-through loop.