feat(engine): reject call activity straight through loop
Call Activities calling itself is one of the scenarios we've seen
causing a straight-through loop. This is only a problem if the call
activity calls it's own process without any wait state in between. E.g.
start event > undefined task > call activity > end event. If the call
activity here calls itself, we're in a straight-through loop.

If there is a wait state in between, e.g.:
start event > user task > call activity > event. We do not detect a
straight-through loop.

This is implemented changing the target element of the call activity.
Instead of checking for the outgoing sequence flows, it checks the start
 event of the called process and continues from here. If it encounters
 the same call activity again (without wait state in between) we know
 we've reached a loop.