fix(atomix): only bind to wildcard interface
When using the SO_BROADCAST flag, Without root privilege, it's
not possible to bind to any address but the wildcard address. It also
does not make sense to support binding to multiple addresses if we're
going to bind to 0.0.0.0 anyway.

Code is simplified to support that, with improved logging, and also
respecting the shutdown time outs configuration.