refactor: replace blacklisting term
Since blacklisting is a general bad term, due to racial aspects, try to replace the term
with a less bad name like: ban. In consequence we need to adjust other places where we use
the old term.

Instead of blacklisting an instance it is now banning an instance or an instance has been banned,
which sounds also a bit better.

We need to verify whether this is really backwards compatible with some benchmarks, currently I think this is the case.