refactor(engine): simplify pending process message subscription state
The previous class hierarchy was a bit confusing. With these changes
here, the `DbProcessMessageSubscriptionState` is responsible for all
state changes. When necessary, it reads from and writes to the transient
state directly.

This also adjusts names and package paths to indicate that reading and
writing transient state is not a mutable operation.