feat(protocol-impl): wrap command distribution records
This adds the ability to a command distribution record to wrap another.
This allows copying of command distribution records and will be used in
an upcoming commit.

The implementation is rather straightforward, setting each of the
individual properties based on the value of the other. Although likely
slower than copying the underlying buffer, this might not be necessary.
This implementation is inspired from
`ProcessInstanceRecord.wrap(ProcessInstanceRecord)`.