fix(transport): avoid concurrent modification of aggregated stream consumers
Makes `AggregatedRemoteStream` thread-safe; stream consumers is now a
`CopyOnWriteArrayList`, which is fine as we expect this to be
read-heavy.

Removes the executor from the `RemoteStreamImpl`, as it was anyway a
different executor than the stream registry's.

Handles a stream's consumer list concurrent modification on each try and
retry. Optimized for the common case, where we pick a random list index
and return the item, at the cost of a slower exception case where
accessing the list throws an exception.