feat: use a separate transaction for pending distributions
As we're moving to running checkers asynchronous (i.e. concurrently) to
the stream processor, we should not be reusing the processing state's
transaction, but instead use a separate transaction.

This can be achieved by using the ScheduledTaskDbState instead of the
ProcessingDbState for such checkers.

In this case, we're switching over the commandredistributor to the
scheduled task state.