feat(raft): replicate complete serialized journal record
To prevent AppendRequest from failing if the journal record schema is different
between the leader and follower, we now send the complete serialized journal record.
This allows the follower to write the bytes directly without serializing them again.
Serializing them again with a different version than the sender can result in a different checksum.