fix: allow to reset logstorage
During running the JMH benchmarks we accumultate a lot of data and we need to free
this at somepoint otherwise we run into issues like OOM or GC slowing us down.

In order to make the logstorage resetable we migrate the entries to a ConcurrentSkipListMap, the
migration itself already gave great performance improvement because of better performance
than a CopyOnWriteList.