test: fix deployment completion wait states
A lot of tests depended on the DeploymentIntent.FULLY_DISTRIBUTED to
know when a deployment is completed. With generalized distribution this
intent is no longer written.

The different way to check for deployment completion is by waiting for a
CommandDistribution.FINISHED event for multi-partition clusters. For a
single partition cluster we need to wait for a Deployment.CREATED event.

Note that we should use .withDistributionIntent to make sure that it is
indeed a Deployment.CREATE command that is distributed.