docs: clarify reuse of the last pending distribution
The if statement might seem out-of-place here. So let's clarify why it
exists.

Looking up the command distribution may be costly. If we have to do this
for every partition we want to distribute it to, then this doesn't scale
so well. We can reuse the last encountered pending distribution if we
encounter it multiple times in a row, to be more efficient.