refactor(engine): add DistributedTypedRecordProcessor
This processor extends the regular TypedRecordProcessor interface.
During processing of a command this interface takes care of checking if
the command was distributed or not. Depending on this is will call one
of two methods.
Processor behavior will differ depending on if the command was already
distributed or not. This interface provides convenience and makes it so
there's no need to check this in every processor individually.