fix: use new state for each checker
Specifically, this allows the engine processors to create a new
ScheduledTaskDbState for each scheduled task, with a new
TransactionContext.

The transaction context is just as important as having a new instance of
the state tree, as both together allow concurrent access to rocksdb.

This is important when one or more scheduled tasks are run async to the
stream processor (concurrently, aka using a different actor).

The change itself is rather simple: don't reuse the ScheduledTaskDbState
between the scheduled tasks, but provide a Factory (or Supplier) to
create new instances for each scheduled task.

The following scheduled tasks are affected:
- DueDateTimerChecker
- MessageTtlChecker
- CommandRedistributor

(cherry picked from commit 368fa3a399093db27ca8e3ecfb8f9160708e6fc2)

Conflicts:
        engine/src/main/java/io/camunda/zeebe/engine/processing/EngineProcessors.java
	engine/src/main/java/io/camunda/zeebe/engine/processing/message/MessageObserver.java