refactor: clarify scheduled task db state factory
We can clarify what the factory really does using JavaDoc on the
interface of the TypedRecordProcessorContext. Of course this requires us
to move that part into the implementation of the context to make sure it
really does align with the documentation.

I specifically chose not to bring the creation of the ProcessingState
into this, because it did not feel right when trying this out. It
would require us to pass a lot of additional parameters into the
constructor which seemed unnecessary. Feel free to make this change if
you think it makes sense in the future.

(cherry picked from commit a9fa5e9e8e83760e245ad3d626f3b20425e6874a)

Conflicts:
  engine/src/main/java/io/camunda/zeebe/engine/Engine.java