refactor: inline processing state construction
I gave it another try (after my previous commit), and it actually works
out pretty nice to move the construction of the processing state to the
TypedRecordProcessorContext as well.

This means that both the ProcessingState and the ScheduledTaskState are
created by the TypedRecordProcessorContextImpl from now on.

This actually simplifies the constructor of the context, and simplifies
the lifecycle of both state objects (i.e. state access trees).