refactor: extract scheduled task state interface
Just like ProcessingState it makes sense to have a ScheduledTaskState.
When using it, we don't need to know that it actually is a using a
DbState.

Originally, this interface was not provided because ScheduledTaskDbState
only comes in an immutable form, while ProcessingDbState has a mutable
and an immutable part of its interface.

Still, using it (and naming its local vars) is easier without the
knowledge about the DB part. Additionally, moving the interface into the
immutable package makes it clear that the state is really read-only.