refactor: extract scheduled task state interface
Just like ProcessingState it makes sense to have a ScheduledTaskState.
When using it, we don't need to know that it actually is a using a
DbState.

Originally, this interface was not provided because ScheduledTaskDbState
only comes in an immutable form, while ProcessingDbState has a mutable
and an immutable part of its interface.

Still, using it (and naming its local vars) is easier without the
knowledge about the DB part. Additionally, moving the interface into the
immutable package makes it clear that the state is really read-only.

(cherry picked from commit 4d3c4e90c995c7483490abe159453fce86fb9604)

Conflicts:
	engine/src/main/java/io/camunda/zeebe/engine/processing/EngineProcessors.java
	engine/src/main/java/io/camunda/zeebe/engine/processing/message/MessageObserver.java
	engine/src/main/java/io/camunda/zeebe/engine/state/ScheduledTaskDbState.java