fix(journal): write records received from older brokers with old sbe schema version
The checksum of the record is calculated from the serialized sbe record which was
written at schema version 1. To ensure that checksum verification does not fail,
serialize the record using old sbe version. Here, we only have to change the version
in the header as there were no changes to the fields in RecordMetadata or RecordData.