feat(test-util): allow limiting by predicate and count
This method allows to limit a stream when a given count of record has
matched the given predicate. This allows us to check, for example, that
two CommandDistribution.FINISHED records have been exported. Since
deployments do a distribution and a resource deletion will also do a
distribution this is especially useful in this case. We can check that
the FINISHED record occurred twice before continuing.