test(engine): fix limiting
This test now uses the new limitByCount method. The limit also changed
slightly as it contained a bug. Before it would check for 3
ResourceDeletion DELETED intents. As we limit this stream to records
written on partition 1 we only get this event once. To be more accurate
we should check for 2 CommandDistribution FINISHED events instead.