fix(protocol): allow serializing UnifiedRecordValue
The CommandDistributionRecord by default contains a UnifiedRecordValue
in it's commandValue property. In some cases this value gets filled
with a proper value, such as a deployment record. However, in a lot of
cases we've decided not to set this property in order to reduce our
message size.

Trying to serialize this empty UnifiedRecordValue to JSON was not
possible. There was no serializer available for this record. By putting
these annotations on the RecordValue interface we can make serialization
 of this record possible.