test: use bpmn model api instead of bpmn files
Generally we prefer using the bpmn model api in tests instead of bpmn
files, because it's much more clear what the test is doing.

Here, I was able to indicate an interesting difference between the
- CreateLargeDeploymentTest: tests that the incoming request's
  maxMessageSize is tested)
- CreateDeploymentTest.shouldRejectDeployIfResourceIsTooLarge: verifies
  that resources that fit the max message size may still exceed the
  batch record size (due to follow-up records).

Those are both integration tests, but we also have an engine test for
this: DeploymentRejectionTest.shouldRejectDeploymentIfResourceIsTooLarge

I'm not sure whether we need both the DeploymentRejectionTest and the
CreateDeploymentTest I mentioned here.